hf_docker/requirements.txt
Flask==2.2.5
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.1.4
Pillow==10.3.0
python-dateutil==2.9.0.post0
pytz==2024.1
typing-extensions==4.12.2
Werkzeug==2.2.3
requests==2.32.3
gdown==5.2.0

# PyTorch CPU-only
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.3.1+cpu
torchvision==0.18.1+cpu
